% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_to_disk_solid_angle.R
\name{disk_to_disk_3d}
\alias{disk_to_disk_3d}
\title{Disk to disk solid angle with 3D perspective view}
\usage{
disk_to_disk_3d(r.source, gap, r.detector, show_plot = TRUE,
  runs = 10000)
}
\arguments{
\item{r.source}{source radius (all units must be consistent)}

\item{gap}{distance between source and detector}

\item{r.detector}{detector radius}

\item{show_plot}{A plot of the results will be displayed by default. Set to
FALSE if this is not desired.}

\item{runs}{Number of particles to simulate. Running more particles improves
accuracy. Default = 1e4.}
}
\description{
Disk to disk solid angle with 3D perspective view
}
\examples{
disk_to_disk_3d(10, 5, 20)

}
