% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_pct_to_num_half_life.R
\name{dk_pct_to_num_half_life}
\alias{dk_pct_to_num_half_life}
\title{Number of half-lives past}
\usage{
dk_pct_to_num_half_life(pct_lost)
}
\arguments{
\item{pct_lost}{Percentage of activity lost since reference time.}
}
\value{
Number of half-lives passed.
}
\description{
Given a percentage reduction in activity, calculate how many
  half-lives have passed.
}
\examples{
dk_pct_to_num_half_life(pct_lost = 93.75)
}
\seealso{
Other decay corrections: \code{\link{dk_correct}},
  \code{\link{dk_time}}
}
\concept{decay corrections}
