% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_pct_to_num_t_h.R
\name{dk_pct_to_num_t_h}
\alias{dk_pct_to_num_t_h}
\title{Number of half-lives past}
\usage{
dk_pct_to_num_t_h(pct_lost)
}
\arguments{
\item{pct_lost}{Percentage of activity lost since reference time.}
}
\value{
Number of half-lives passed.
}
\description{
Given a percentage reduction in activity, calculate how many
  half-lives have passed.
}
\examples{
dk_pct_to_num_t_h(93.75)
}
