% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnp_si_sp_hist.R
\name{mcnp_si_sp_hist}
\alias{mcnp_si_sp_hist}
\title{energy distribution histogram entries}
\usage{
mcnp_si_sp_hist(fun_mode = 1, bound_mode, emin, emax, bin_prob)
}
\arguments{
\item{fun_mode}{1 (default) or 2. Run function with copy and paste (1) or with data loaded in R (2).}

\item{bound_mode}{1 for "lower" or 2 for "upper". Identify whether bin energy data is lower (minimum energy of each bin) or upper (maximum energy of each bin). If the length of the energy vector is 1 greater than the number of probabilities (bin_prob), then the highest (for case 1) or lowest (for case 2) value will be the maximum (case 1) or minimum (case 2) bounding energy. Otherwise, if the vector lengths are equal, a maximum or minimum bounding energy will be needed.}

\item{emin}{Lower bounding energy. (See bound_mode)}

\item{emax}{Upper bounding energy. (See bound_mode)}

\item{bin_prob}{A vector of the bin probabilities.}
}
\value{
A vector of energy bins and probabilities for an energy distribution, formatted as needed for MCNP input. It is designed for copy and paste into an MCNP input. (The # should be changed to the appropriate distribution number. The data is saved in the global environment and appended to a file in the user's working directory, si_sp.txt. Two plots of the data are provided to the plot window,  one with two linear axes and one with two log axes.
}
\description{
Make MCNP histogram probabilities for energy bins.
}
\details{
In copy-and-paste mode (fun_mode = 1), the user will copy and paste, when prompted, a vector of minimum or maximum bin energies, a vector of probabilities for the bins, and a single bounding upper or bin energy, if needed. Paste values may be copied and pasted from spreadsheets or text files, may be tab-separated or space-separated and may not include commas. In data-loaded mode (fun_mode = 2), the data may be identified by named vector, e.g. my_emin_data, or by column of a data frame, e.g. photons_cs137_hist[1] as a emax example and photons_cs137_hist[2] as a bin_prob example.
}
\examples{
mcnp_si_sp_hist(fun_mode = 2, 
bound_mode = 2, 
emin = 0, 
emax = photons_cs137_hist[1:64, 1], 
bin_prob = photons_cs137_hist[1:64, 2])
}
\seealso{
Other mcnp tools: \code{\link{mcnp_cone_angle}},
  \code{\link{mcnp_est_nps}},
  \code{\link{mcnp_matrix_rotations}},
  \code{\link{mcnp_plot_out_spec}},
  \code{\link{mcnp_scan2spec}}, \code{\link{mcnp_si_sp_RD}}
}
\concept{mcnp tools}
