% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnp_si_sp_RD.R
\name{mcnp_si_sp_RD}
\alias{mcnp_si_sp_RD}
\title{Produce MCNP source terms from ICRP 107 data except beta}
\usage{
mcnp_si_sp_RD(desired_RN, rad_type = NULL, photon = FALSE, cut = 0,
  erg.dist = 1)
}
\arguments{
\item{desired_RN}{Radionuclide in form Ba-137m}

\item{rad_type}{Radiation type, leave NULL if selecting photons or
select from:
'X' for X-Ray
'G' for Gamma
'AE' for Auger Electron
'IE' for Internal Conversion Electron
'A' for Alpha
'AR' for Alpha Recoil
'B-' for Beta Negative
'AQ' for Annihilation Quanta
'B+' for Beta Positive
'PG' for Prompt Gamma
'DG' for Delayed Gamma
'DB' for Delayed Beta
'FF' for Fission Fragment
'N' for Neutron}

\item{photon}{'Y' to select all rad_types that are photons}

\item{cut}{minimum energy defaults to 0}

\item{erg.dist}{energy distribution number for MCNP input}
}
\value{
a data frame can be saved to memory if desired (i.e. by my_file <- mcnp_si_sp_RD(...)).
For use with MCNP, a text file, 'si.sp.dat' is written to working directory.
If file already exists, it is appended. The file contains all
emission energies in the si 'card' and the Line indicator, L is included,
e.g. si1  L  0.01 (showing a first energy of 0.01 MeV).
This is followed by the emission probability of each si entry.
An additional text entry is made summing up the probabilities.
}
\description{
Obtain emission data from the RadData package and write to a file for use with the radiation transport code, MCNP.
}
\examples{
mcnp_si_sp_RD("Co-60", photon = TRUE, cut = 0.01, erg.dist = 13)
mcnp_si_sp_RD("Sr-90", rad_type = "B-", cut = 0.01, erg.dist = 15)
mcnp_si_sp_RD("Am-241", rad_type = "A", cut = 0.01, erg.dist = 23)
}
\seealso{
[si_hist()] and [sp_hist()]  if radioactive emission data is available in histogram form and needs formatting for MCNP input.
[RN_screen_plot] may be used for a simple plot of output.

Other mcnp tools: \code{\link{mcnp_cone_angle}},
  \code{\link{mcnp_est_nps}},
  \code{\link{mcnp_matrix_rotations}},
  \code{\link{mcnp_plot_out_spec}},
  \code{\link{mcnp_scan2spec}},
  \code{\link{mcnp_si_sp_hist_scan}},
  \code{\link{mcnp_si_sp_hist}}
}
\concept{mcnp tools}
