% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnp_sp_hist.R
\name{mcnp_sp_hist}
\alias{mcnp_sp_hist}
\title{energy distribution histogram entries}
\usage{
mcnp_sp_hist(bin_prob)
}
\arguments{
\item{bin_prob}{A vector of the bin probabilities.}
}
\value{
A matrix of values for source probabilities to copy and paste into an MCNP input. (The # should be changed to the appropriate distribution number. The NA's in the last row should be discarded.)
}
\description{
Make MCNP histogram probabilities for energy bins.
}
\examples{
mcnp_sp_hist(rep(1 / 11, 11))
}
\seealso{
[mcnp_si_hist()]

Other mcnp tools: \code{\link{mcnp_cone_angle}},
  \code{\link{mcnp_est_nps}},
  \code{\link{mcnp_matrix_rotations}},
  \code{\link{mcnp_plot_out_spec}},
  \code{\link{mcnp_scan2spec}}, \code{\link{mcnp_si_hist}},
  \code{\link{mcnp_si_sp_RD}}
}
\concept{mcnp tools}
