% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photons_cs137_hist.R
\docType{data}
\name{photons_cs137_hist}
\alias{photons_cs137_hist}
\title{File Description:}
\format{
A \code{data.frame}
\describe{
\item{E_max}{Maximum Energy in MeV}
\item{bin_tally}{Tally result for this bin}
\item{R}{Monte Carlo uncertainty for this bin}
}
}
\usage{
photons_cs137_hist
}
\description{
This data file was generated in MCNP from a model of Gamma Well Irradiator
with no attenuator in place. MCNP will include in the output a histogram of
tally results when there is an E Tally Energy card. Results in the output up
to MCNP version 6 have no headers, but the columns are:
}
\keyword{datasets}
