% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spec_from_hist.R
\name{plot_spec_from_hist}
\alias{plot_spec_from_hist}
\title{Convert histogram data to average points and plot as spectrum.}
\usage{
plot_spec_from_hist(spec.df, title)
}
\arguments{
\item{spec.df}{A data frame with no header. Maximum energy in MeV should be
in the first column, binned results in the second column, uncertainty in
the third column.}

\item{title}{Title for chart}
}
\description{
Model results from MCNP and perhaps other sources typically
  provide binned tally results with columns representing maximum energy in
  MeV, a column with the mean tally result titled "mean" and an uncertainty
  column titled "R".
}
\examples{
library(ggplot2)
plot_spec_from_hist(photons_cs137_hist, "example Cs-137 well irradiator")
}
