% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsafer.R
\docType{package}
\name{radsafer}
\alias{radsafer}
\alias{radsafer-package}
\title{radsafer package}
\description{
Functions for Radiation Safety
}
\author{
Mark Hogue \email{mark.hogue.chp@gmail.com}
}
