% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RN_bin_screen_plot.R
\name{RN_bin_screen_plot}
\alias{RN_bin_screen_plot}
\title{Quick plot to accompany [RN_bin_screen_phot]}
\usage{
RN_bin_screen_plot(df)
}
\arguments{
\item{df}{data frame of results including RN (radionuclide), 
energy in E_MeV and probability (prob) of photon.

@examples 
RN_bin_screen_plot(spec_0.1_0.3)

@export}
}
\description{
Plots with ggplot2 with geom_point and log y-scale.
Useful for a small set of radionclides.
The point of this is to make it easy to get the plot
by entering only the data frame name.
}
\seealso{
Other radionuclides: \code{\link{RN_Spec_Act}},
  \code{\link{RN_bin_screen_phot}}, \code{\link{RN_info}},
  \code{\link{RN_screen}}, \code{\link{search_alpha_by_E}},
  \code{\link{search_beta_by_E}},
  \code{\link{search_phot_by_E}}
}
\concept{radionuclides}
