% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RN_find_parent.R
\name{RN_find_parent}
\alias{RN_find_parent}
\title{Find a potential precursor of a radionuclide
 @description Find a potential parent radionuclide by searching the progeny fields in RadData ICRP_07.NDX}
\usage{
RN_find_parent(RN_select)
}
\arguments{
\item{RN_select}{identify the radionuclide of interest in the format "Es-254m"}
}
\value{
a subset of the data frame RadData::ICRP_07.NDX
}
\description{
Find a potential precursor of a radionuclide
 @description Find a potential parent radionuclide by searching the progeny fields in RadData ICRP_07.NDX
}
\examples{
RN_find_parent("Th-230")
RN_find_parent("Tl-208")
}
