% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RN_quick_table.R
\name{RN_info}
\alias{RN_info}
\title{Quick table of Radionuclide Data from the RadData package}
\usage{
RN_info(RN_select)
}
\arguments{
\item{RN_select}{identify the radionuclide of interest in the format "Es-254m"}
}
\value{
a table including half-life, decay modes, decay progeny, and branch fractions
}
\description{
Access a quick summary of radionuclide data. This is for convenience only and does not replace a more comprehensive view as is available in the Radiological Toolbox <doi:10.2172/1201298>
}
\examples{
RN_info("Es-254m")
RN_info("Cf-252")
RN_info("Cs-137")
RN_info("Am-241")
}
\seealso{
Other radionuclides: \code{\link{RN_Spec_Act}},
  \code{\link{RN_bin_screen_phot}},
  \code{\link{RN_bin_screen_plot}},
  \code{\link{RN_screen}}, \code{\link{search_alpha_by_E}},
  \code{\link{search_beta_by_E}},
  \code{\link{search_phot_by_E}}
}
\concept{radionuclides}
