% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RN_plt.R
\name{RN_plt}
\alias{RN_plt}
\title{Quick plot to radionuclide emission data screens.}
\usage{
RN_plt(df)
}
\arguments{
\item{df}{data frame of results including RN (radionuclide),
energy in E_MeV and probability (prob) of photon.}
}
\value{
plot of spectrum
}
\description{
Plots with ggplot2 with geom_point and log y-scale.
Useful for a small set of radionclides.
The point of this is to make it easy to get the plot
by entering only the data frame name.
}
\examples{
RN_plt(spec_0.1_0.3)
}
\seealso{
Other radionuclides: \code{\link{RN_Spec_Act}},
  \code{\link{RN_index_screen}}, \code{\link{RN_info}},
  \code{\link{bin_screen_phot}},
  \code{\link{search_alpha_by_E}},
  \code{\link{search_beta_by_E}},
  \code{\link{search_phot_by_E}}
}
\concept{radionuclides}
