% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnp_rot_transform.R
\name{rot_fun}
\alias{rot_fun}
\title{MCNP coordinate transformation rotation}
\usage{
rot_fun(rot.axis, angle_degrees)
}
\arguments{
\item{rot.axis}{axis (x, y or z) with respect to which the rotation is to be made.}

\item{angle_degrees}{angle, in degrees, of rotation about the rotation axis}
}
\value{
A transformation matrix to include with a transformation card.
}
\description{
Make rotation matrix for use with MCNP transformation card. For combined rotations, use matrix multiplication (%*%)
}
\examples{
rot_fun("x", 30)
rot_fun("y", 7)
rot_fun("z", 15)
rot_fun("x", 45) \%*\% rot_fun("y", 45)
}
