% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler_sim.R
\name{scaler_sim}
\alias{scaler_sim}
\title{Count Room Scaler Simulation}
\usage{
scaler_sim(true_bkg, true_samp, runs, trials)
}
\arguments{
\item{true_bkg}{True background count rate in counts per minute.}

\item{true_samp}{True sample count rate in counts per minute.}

\item{runs}{How many iterations of counting are performed.}

\item{trials}{How many times to run the model.}
}
\value{
A histogram of all trial results including limits for +/- 1 standard
  deviation.
}
\description{
Returns a plotted distribution of results for a scaler model
  based on the Poisson distribution. Inputs and outputs in counts per minute.
}
\examples{
scaler_sim(true_bkg = 50, true_samp = 10, runs = c(1, 10, 50, 100), trials = 1e5)
}
