% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnp_scan2spec.df.R
\name{scan2spec.df}
\alias{scan2spec.df}
\title{Copy and paste MCNP output spectral data for use with `mcnp_plot_out_spec.R`}
\usage{
scan2spec.df()
}
\value{
spectrum file with maximum energy and MCNP bin value
}
\description{
Provides quick copy-and-paste conversion to data frame.
Converion is based on the three columns with 
results from MCNP output in energy bins with
maximum energy in MeV in the first column,
the mean result for the bin in the second column, and 
relative Monte Carlo uncertainty in the third column.
}
