% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_alpha_by_E.R
\name{search_alpha_by_E}
\alias{search_alpha_by_E}
\title{Search for alpha}
\usage{
search_alpha_by_E(E_min = 0, E_max = 10,
  min_half_life_seconds = NULL, max_half_life_seconds = NULL,
  min_prob = 0)
}
\arguments{
\item{E_min}{minimum energy in MeV, default = 0}

\item{E_max}{maximum energy in MeV, default = 10}

\item{min_half_life_seconds}{minimum half-life in seconds.
Use multiplier as needed, e.g. 3 * 3600 for 3 hours. Default = NULL,}

\item{max_half_life_seconds}{maximum half-life. See min_half_life_seconds.}

\item{min_prob}{minimum probability with default = 0.}
}
\value{
search results in order of half-life. Recommend assigning
results to a viewable object, such as 'search_results'
}
\description{
Search for alpha emission based on energy, half-life and minimum probability.
}
\examples{
# between 7 and 8 MeV
search_results <- search_alpha_by_E(7, 8)

# 1-4 MeV; half-life between 1 and 4 hours
search_results <- search_alpha_by_E(1, 4, 1 * 3600, 4 * 3600)

# between 7 and 10 MeV with at least 1e-3 probability
search_results <- search_alpha_by_E(7, 10, min_prob = 1e-3)
}
\seealso{
[RN_plt()]

Other radionuclides: \code{\link{RN_Spec_Act}},
  \code{\link{RN_index_screen}}, \code{\link{RN_info}},
  \code{\link{RN_plot_spectrum}}, \code{\link{RN_plt}},
  \code{\link{bin_screen_phot}},
  \code{\link{search_beta_by_E}},
  \code{\link{search_phot_by_E}}
}
\concept{radionuclides}
