% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnp_energy_hist.R
\name{si_hist}
\alias{si_hist}
\title{MCNP histograms source energies}
\usage{
si_hist(emin, emax)
}
\arguments{
\item{emin}{A vector of energies forming the lower bound of the bin. (emin values other than the first value also provide a bin emax.)}

\item{emax}{A single energy with the upper bound of the highest bin.}
}
\value{
A matrix of values for source input to copy and paste into an MCNP input. (The # should be changed to the appropriate distribution number. The NA's in the last row should be discarded.)
}
\description{
Make MCNP histogram energy bins cards for source definition.
}
\examples{
si_hist(1:10 / 10, 1.2)
}
