% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnp_energy_hist.R
\name{sp_hist}
\alias{sp_hist}
\title{MCNP histograms source bin probabilities}
\usage{
sp_hist(bin_prob)
}
\arguments{
\item{bin_prob}{A vector of the bin probabilities.}
}
\value{
A matrix of values for source probabilities to copy and paste into an MCNP input. (The # should be changed to the appropriate distribution number. The NA's in the last row should be discarded.)
}
\description{
Make MCNP histogram probabilities for energy bins.
}
\examples{
sp_hist(rep(1 / 11, 11))
}
