% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_0.1_0.3.R
\docType{data}
\name{spec_0.1_0.3}
\alias{spec_0.1_0.3}
\title{Output from example in [RN_bin_screen_phot]}
\format{A \code{data.frame}
\describe{
\item{RN}{Radionuclide name}
\item{code_AN}{radiation type alpha-numeric}
\item{E_MeV}{photon energy in mega-electronvolts}
\item{prob}{probability of this photon per transformation}
\item{code_num}{radiation type number}
\item{is_photon}{TRUE if photon}
}}
\usage{
spec_0.1_0.3
}
\description{
Available for follow-up example in [RN_bin_screen_plot]
}
\keyword{datasets}
