% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stay_time.R
\name{stay_time}
\alias{stay_time}
\title{Stay time for radiation work.}
\usage{
stay_time(dose_rate, dose_allowed, margin = 20)
}
\arguments{
\item{dose_rate}{Dose rate per hour for the work - units consistent with dose allowance, e.g. mRem/h, microSv/h.}

\item{dose_allowed}{Dose that can not be exceeded for this job.}

\item{margin}{Percent margin to protect limit, default = 20 percent.}
}
\value{
Time in minutes allowed for the work.
}
\description{
Calculate stay time for radiation work.
}
\examples{
stay_time(dose_rate = 100, dose_allowed = 50, margin = 20)
}
