% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnp_cone_angle.R
\name{tan_d2}
\alias{tan_d2}
\title{MCNP Cone Opening Parameter}
\usage{
tan_d2(d)
}
\arguments{
\item{d}{The cone angle in degrees.}
}
\value{
The ratio of actual to reference air density.
}
\description{
MCNP cone surface requires a term, t^2, which is the tangent of
  the cone angle, in radians, squared. This function takes an input in
  degrees and provides the parameter needed by MCNP.
}
\examples{
tan_d2(45)
}
