% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_estimate.R
\name{tau_est}
\alias{tau_est}
\title{Estimate tau parameter for ratemeter_sim}
\usage{
tau_est(pct_eq, t_eq)
}
\arguments{
\item{pct_eq}{Percent equilibrium}

\item{t_eq}{Time, in seconds, to the given percent equilibrium is achieved.}
}
\value{
tau, the time constant, in seconds.
}
\description{
If the time constant is not known, but the vendor specifies that
  the ratemeter will reach some percentage of equilibrium in some number of
  seconds, use this function to estimate tau.
}
\examples{
tau_est(90, 22)
}
