% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_activity.R
\name{dk_activity}
\alias{dk_activity}
\title{Time for a radionuclide to decay to a target activity.}
\usage{
dk_activity(A0, half_life, t)
}
\arguments{
\item{A0}{The original activity, or related parameter.}

\item{half_life, }{Half-life. Units are arbitrary, but must match time past.}

\item{t}{The time past since original activity.}
}
\value{
Time, in same units as half-life, to decay to target activity.
}
\description{
Calculate time for a radionuclide to decay to a target activity.
}
\examples{
dk_activity(10, 8, 60)
}
