% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_time.R
\name{dk_time}
\alias{dk_time}
\title{Time to decay to target radioactivity.}
\usage{
dk_time(half_life, A0, A1)
}
\arguments{
\item{half_life, }{Half-life. Units are arbitrary, but must match time past.}

\item{A0}{The original activity, or related parameter.}

\item{A1}{The target activity.}
}
\value{
Time, in same units as half-life, to decay to target activity.
}
\description{
Calculate time for a radionuclide to decay to a target activity.
}
\examples{
dk_time(5770, 200, 1)
}
