% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnp_plot_out_spec.R
\name{plot_spec_from_hist}
\alias{plot_spec_from_hist}
\title{Convert histogram data to average points and plot as spectrum.}
\usage{
plot_spec_from_hist(spec.df, units, title)
}
\arguments{
\item{spec.df}{A data frame with no header. Maximum energy in MeV should be in the first column, binned results in the second column, uncertainty (optionally) in the third column.}

\item{units}{Result units}

\item{title}{Title for chart}
}
\description{
Model results from MCNP in energy bins with   maximum energy in MeV in the first column and the mean result for the bin in the second column.
}
\examples{
library(ggplot2)
plot_spec_from_hist(photons_cs137_hist[1:16, ], "mR_h", "Example: Cs-137 with scatter")
}
