% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RN_Spec_Act.R
\name{RN_Spec_Act}
\alias{RN_Spec_Act}
\title{Specific Activity}
\usage{
RN_Spec_Act(RN_select)
}
\arguments{
\item{RN_select}{identify the radionuclide of interest in the format "Es-254m"}
}
\value{
specific activity in Bq / g
}
\description{
Provides specific activity of a radionuclide in Bq/g.
}
\examples{
RN_Spec_Act("Ac-230")
RN_Spec_Act("At-219")
RN_Spec_Act("Es-251")
RN_Spec_Act("Pd-96")
RN_Spec_Act("Te-117")
RN_Spec_Act("Ba-137m")
}
\seealso{
Other radionuclides: \code{\link{RN_index_screen}},
  \code{\link{RN_info}},
  \code{\link{RN_plot_search_results}},
  \code{\link{RN_plot_spectrum}},
  \code{\link{RN_save_spectrum}},
  \code{\link{RN_search_alpha_by_E}},
  \code{\link{RN_search_beta_by_E}},
  \code{\link{RN_search_phot_by_E}},
  \code{\link{bin_screen_phot}},
  \code{\link{search_alpha_by_E}},
  \code{\link{search_beta_by_E}},
  \code{\link{search_phot_by_E}}
}
\concept{radionuclides}
